Fixes to artificial landscapes and very large cell size
Artificial landscape generator modified to prevent skewed distribution of cell values in fractal continuous landscapes - new recursive function implemented.
Computations of density etc. altered to avoid integer overflow if cell size is very large (e.g. 50000m) - all calculations now performed as float or double.

+++ src/Common/BatchMode.cpp
@@ -6811,7 +6811,7 @@ else {
 	for (int j = 0; j < dem.nSeasons; j++) {
 		for (int i = 0; i < paramsLand.nHabMax; i++) {
 			parameters >> k;
-			k *= ((double)(paramsLand.resol*paramsLand.resol))/10000.0;
+			k *= (((float)paramsLand.resol*(float)paramsLand.resol))/10000.0f;
 				pSpecies->setHabK(i,j,k);			
 		}
 	}
